backend/requirements.txt
fastapi==0.109.0
uvicorn[standard]==0.27.0
python-jose[cryptography]==3.3.0
httpx==0.26.0
pydantic==2.5.3
pydantic-settings==2.1.0
python-multipart==0.0.6
asyncpg==0.29.0
sqlalchemy[asyncio]==2.0.25
alembic==1.13.1
redis==5.0.1
python-dotenv==1.0.0
